move function to utils from types
